# Optimized `ByteUtil.Hex()`