Should be about 8 times faster. Tested with various lengths and sizes.

|   Method |      Mean |     Error |    StdDev |
|--------- |----------:|----------:|----------:|
| Original | 632.75 us | 12.454 us | 23.083 us |
|   Lookup |  78.23 us |  1.520 us |  2.029 us |